+++ changelog.md
@@ -4,7 +4,7 @@
 
 - Renamed `lib` table to `package`, But you can still use `lib` as cherry is backward-compatible!
 
-> NOTE 1: Compatibility was made from this version so it can run packages that were created with old versions of cherry.
+> NOTE: Compatibility was made from this version so it can run packages that were created with old versions of cherry.
 
 ### v0.2 (8/October/2020)