+++ changelog.md
@@ -3,6 +3,7 @@
 ### v0.3 (9/October/2020)
 
 - Renamed `lib` table to `package`, But you can still use `lib` as cherry is backward-compatible!
+- Added more debug info when running commands!
 
 > NOTE: Compatibility was made from this version so it can run packages that were created with old versions of cherry.